--- apps/igloo-kit-cli/src/infrastructure/setup/mod.rs
@@ -1,21 +1,0 @@
-pub mod network;
-pub mod scaffold;
-pub mod container;
-pub mod validate;
-pub mod database;
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
